2. Парковка машин [#853 #255239]

+++ src/main/java/ru/job4j/ood/lsp/parking/Car.java
@@ -1,5 +1,5 @@
 package ru.job4j.ood.lsp.parking;
 
 public interface Car {
-    public double getMass();
+    public int getMass();
 }
